apps/scanner/requirements.txt
opencv-python>=4.9
aiohttp>=3.9
supabase>=2.0
rich>=13.0
numpy>=1.26
Pillow>=10.0
python-dotenv>=1.0
pydantic>=2.0
pytest>=8.0
pytest-asyncio>=0.23
aioresponses>=0.7
